% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_rf.R
\name{do_rf}
\alias{do_rf}
\title{Collect all data from store, prepare data, train random forest}
\usage{
do_rf(
  config,
  store,
  n_trees = NULL,
  n_samples = NULL,
  prepare_data_only = TRUE
)
}
\arguments{
\item{config}{in list form from \code{\link[yaml:read_yaml]{yaml::read_yaml()}}}

\item{store}{containing all the data}

\item{n_trees}{Priority order is: argument \code{n_trees} > \code{config$n_trees} > default \code{500}}

\item{n_samples}{Priority order is: argument \code{n_samples} > \code{config$n_samples} > default \code{50}}

\item{prepare_data_only}{only prepare the data and then return}
}
\value{
list with prepared data or model
}
\description{
Collect all data from store, prepare data, train random forest
}
\examples{
# TODO
}
\seealso{
\itemize{
\item \code{\link[rmweather:rmw_do_all]{rmweather::rmw_do_all()}}
}
}
