% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_rf.R
\name{do_rf_list}
\alias{do_rf_list}
\title{Runs a list of configs}
\usage{
do_rf_list(list_of_configs, store, description, ...)
}
\arguments{
\item{list_of_configs}{a list of configs to run}

\item{store}{a local store containing all the data needed for the run}

\item{description}{of the models. Will be saved in statistics.yml and a run file with the stats from all models
with sanitized desription as file name is created in \code{output_dir}.}

\item{...}{
  Arguments passed on to \code{\link[=do_rf]{do_rf}}
  \describe{
    \item{\code{config}}{in list form from \code{\link[yaml:read_yaml]{yaml::read_yaml()}}}
    \item{\code{n_trees}}{Priority order is: argument \code{n_trees} > \code{config$n_trees} > default \code{500}}
    \item{\code{n_samples}}{Priority order is: argument \code{n_samples} > \code{config$n_samples} > default \code{50}}
    \item{\code{prepare_data_only}}{only prepare the data and then return}
    \item{\code{output_dir}}{if not NULL all data will be saved under \verb{output_dir\\config$site\\xxhash64} (xxhash64 is calculated
from the return value from \code{\link[rmweather:rmw_do_all]{rmweather::rmw_do_all()}} with \code{\link[digest:digest]{digest::digest()}})}
  }}
}
\value{
list of results from do_rf
}
\description{
Runs a list of configs
}
\seealso{
\code{\link[rmweather:rmw_do_all]{rmweather::rmw_do_all()}}, \code{\link[=do_rf]{do_rf()}}
}
