% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_from_remote.R
\name{download_from_remote}
\alias{download_from_remote}
\title{Download the necessary data from a remote store in a local}
\usage{
download_from_remote(local, remote, config)
}
\arguments{
\item{local}{local rOstluft store. Probably a \code{\link[rOstluft:storage_local_rds]{rOstluft::storage_local_rds()}}}

\item{remote}{remote rOstluft store. Probably a \code{\link[rOstluft:storage_s3_rds]{rOstluft::storage_s3_rds()}}}

\item{config}{config in list form from \code{\link[yaml:read_yaml]{yaml::read_yaml()}}}
}
\value{

}
\description{
Download the necessary data from a remote store in a local
}
\examples{
config <- get_example_config()
remote <- rOstluft::store_aqmet()
local <- rOstluft::storage_local_rds(
  "rf_example",
  format = rOstluft::format_rolf(),
  read.only = FALSE
)

download_from_remote(local, remote, config)

# tidy up
local$destroy("DELETE")
}
