% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_example_config}
\alias{get_example_config}
\title{Reads example yaml from disk}
\usage{
get_example_config()
}
\value{
output from\code{\link[yaml:read_yaml]{yaml::read_yaml()}}
}
\description{
Reads example yaml from disk
}
\examples{
# config file content
fn <- system.file("extdata", "rf_config.yml", package = "rOstluft.rmweather", mustWork = TRUE)
cat(readr::read_file(fn))


# get the parsed yaml file
get_example_config()
}
