% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_rf.R
\name{prepare_observations_rf}
\alias{prepare_observations_rf}
\title{Get data from store and rearange to wide format}
\usage{
prepare_observations_rf(config, store)
}
\arguments{
\item{config}{in list form from \code{\link[yaml:read_yaml]{yaml::read_yaml()}}}

\item{store}{containing all the data}
}
\value{
list with items observations and n_observations as summary
}
\description{
Get data from store and rearange to wide format
}
\examples{
config <- get_example_config()
store <- rOstluft::store_aqmet()
prepare_observations_rf(config, store)
}
