% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOstluft.rmweather-package.R
\docType{package}
\name{rOstluft.rmweather-package}
\alias{rOstluft.rmweather}
\alias{rOstluft.rmweather-package}
\title{rOstluft.rmweather: Meteorological normalisation with rmweather with data from rOstluft stores}
\description{
This package creates wrapper around rmweather to create meteorological normalisation
  using Random Forest. The role of rOstluft is to provide the input data needed.
}
\author{
\strong{Maintainer}: Thomas von Allmen \email{thomas.vonallmen@zuerich.ch}

Authors:
\itemize{
  \item Jörg Sintermann
}

Other contributors:
\itemize{
  \item Ostluft [copyright holder]
}

}
\keyword{internal}
